refactor: accordion src
