+++ packages/accordion/src/index.ts
@@ -1,2 +1,6 @@
-export * from "./accordion"
+export { Accordion, type AccordionProps } from "./accordion"
+export { AccordionButton, type AccordionButtonProps } from "./accordion-button"
+export { AccordionIcon } from "./accordion-icon"
+export { AccordionItem, type AccordionItemProps } from "./accordion-item"
+export { AccordionPanel, type AccordionPanelProps } from "./accordion-panel"
 export * from "./use-accordion"
